Update config path for neovim
Neovim follows xdg style directories which means that all files
live under ~/.config/nvim.  When the generator was switches to go
the assumptions was that the config files were named the same as
vim but just under .config.  This assumption is incorrect.

This restores the same functionality and names as the previous
python generator.

  Basedir      : ~/.config/nvim
  Main rc file : ~/.config/nvim/init.vim
  Local rc file: ~/.config/nvim/local_init.vim
  Local bundle : ~/.config/nvim/local_bundles.vim